# docs/openapi/auth-internal.json
{
  "openapi": "3.1.0",
  "info": {"title": "SSO Lab Auth Server Internal Admin API", "version": "1.0.0"},
  "servers": [{"url": "http://auth-server:8080"}],
  "security": [{"internalClient": []}],
  "paths": {
...
    "/internal/admin/v1/users/{userId}": {"get": {"responses": {"200": {"description": "Masked identity detail"}, "404": {"description": "Not found"}}}},
...
    "/internal/admin/v1/users/{userId}/email/reveal": {"post": {"responses": {"200": {"description": "Decrypted email after fresh Admin re-authentication"}, "401": {"description": "Re-authentication required"}}}},
...
    "/internal/admin/v1/reauth/verify": {"post": {"responses": {"200": {"description": "Short-lived Admin re-authentication proof"}, "400": {"description": "Generic verification failure"}}}},
...
  },
  "components": {"securitySchemes": {"internalClient": {"type": "apiKey", "in": "header", "name": "X-Internal-Secret"}}}
}